better organizing the mapper structure
