+++ frontend-v2-sarah/src/utils/date.ts
@@ -1,0 +1,3 @@
+export const timestampMsFormatter = (isoDate: string): number => {
+  return new Date(isoDate).getTime();
+};
